restructure folder for pages
