+++ src/shared/app.tsx
@@ -6,9 +6,7 @@ import { styled } from "onefx/lib/styletron-react";
 import React, { Component } from "react";
 import { Switch } from "react-router";
 import { Route } from "react-router-dom";
-import { ActionDetail } from "./action/action-detail";
 import { Actions } from "./actions/actions";
-import { AddressDetailsPage } from "./address-details/address-details-page";
 import { BlockDetail } from "./block/block-detail";
 import { Blocks } from "./block/blocks";
 // @ts-ignore
@@ -20,6 +18,8 @@ import { colors } from "./common/styles/style-color";
 import { fonts } from "./common/styles/style-font";
 import { TOP_BAR_HEIGHT, TopMenuBar } from "./common/top-menu-bar";
 import { Home } from "./home/home";
+import { ActionDetailPage } from "./pages/action-detail-page";
+import { AddressDetailsPage } from "./pages/address-details-page";
 import { Wallet } from "./wallet/wallet";
 
 type Props = {
@@ -54,7 +54,7 @@ export class App extends Component<Props> {
               />
               <Route path="/block/:height" component={BlockDetail} />
               <Route exact path="/block" component={Blocks} />
-              <Route exact path="/action/:hash" component={ActionDetail} />
+              <Route exact path="/action/:hash" component={ActionDetailPage} />
               <Route exact path="/action" component={Actions} />
               <Route path="/wallet" component={Wallet} />
               <Route component={NotFound} />
